+++ kbl/kbl-v25/src/main/resources/kbl25/KBL25_SR1-ext.xjb
@@ -73,7 +73,7 @@
 
     <jxb:bindings multiple="true"
             node="//xs:complexType[@name!='Alias_identification' and .//xs:element[@name='Alias_id']]">
-        <inheritance:implements>com.foursoft.harness.kbl.common.HasAliasId&lt;com.foursoft.harness.kbl.v25.KblAliasIdentification&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.harness.kbl.v25.HasAliasIds</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Description']]">
